% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blade.R
\name{blade}
\alias{blade}
\title{BLADE main function wrapper}
\usage{
blade(Y, mu, sigma, a, k0, a0)
}
\arguments{
\item{Y}{description to add}

\item{mu}{description to add}

\item{a}{description to add}

\item{k0}{description to add}

\item{a0}{description to add}
}
\description{
call elbo1 to elbo5 functions to perform variational inference
}
