% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cors_df}
\alias{cors_df}
\title{Dataset of wire and blade alignment cross-correlation (ccf)}
\format{
A tibble objects with 3 variables:
\itemize{
\item lag: The lag of the alignment position.
\item ccf: The cross-correlation of the aligned position.
\item ns: The number of observations the ccf is based on.
}
}
\usage{
cors_df
}
\description{
This dataset is the alignment cross-correlation of \code{x3p} wire and blade cuts.
}
\keyword{datasets}
