% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_topk.R
\name{df_topk}
\alias{df_topk}
\title{Top k clustered local maximums}
\usage{
df_topk(data, k = 10, probs = 0.5, len_block = 100)
}
\arguments{
\item{data}{A alignment data frame with two columns: \code{lag} and \code{ccf}.}

\item{k}{A positive integer to control the number of resulting local maximums.}

\item{probs}{A value between 0 and 1 to control the data used for clustering.}

\item{len_block}{A positive number controlling the size of neighborhood.}
}
\value{
A data frame with three columns:
\itemize{
\item .cluster: The cluster id.
\item lag: The lag in the alignment.
\item ccf: The cross-correlation value in the alignment.
}
}
\description{
This function uses density-based spatial clustering of applications with noise (DBSCAN) to get top k clustered local maximum from an alignment data frame.
}
