% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxcorrTest.R
\name{maxcorrTest}
\alias{maxcorrTest}
\title{Test of the maximum correlation}
\usage{
maxcorrTest(sample, block, alpha = 0.05)
}
\arguments{
\item{sample}{vector of successive cross-correlation values}

\item{block}{integer value of the suggested block size. Smaller blocks will
create larger number of local maxima but increase possible dependence.}

\item{alpha}{significance value/acceptable level of Type 1 error.}
}
\value{
object of class \code{htest} (hypothesis test)
}
\description{
Test of the maximum correlation
}
\examples{
# for the example data both a test for the max and a test for the min
# CCF result in the expected outcome:
maxcorrTest(cors_df$ccf, 100) # highly significant
maxcorrTest(-cors_df$ccf, 100)  # not significant
}
