% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_topk.R
\name{df_topk}
\alias{df_topk}
\title{Top k clustered local maximums}
\usage{
df_topk(data, k = 10, len_block = 100)
}
\arguments{
\item{data}{A alignment data frame with two columns: \code{lag} and \code{ccf}.}

\item{k}{A positive integer to control the number of resulting local maximums.}

\item{len_block}{A positive number controlling the size of neighborhood.}
}
\value{
A data frame with three columns:
\itemize{
\item .cluster: The cluster id.
\item lag: The lag in the alignment.
\item ccf: The cross-correlation value in the alignment.
}
}
\description{
This function uses density-based spatial clustering of applications with noise (DBSCAN) to get top k clustered local maximum from an alignment data frame.
}
\examples{
library(ggplot2)
topk <- df_topk(cors_df)
cors_df \%>\%
  ggplot(aes(x = lag, y = ccf)) +
  geom_point() +
  geom_point(aes(color = "topk"), data = topk) +
  theme_bw()

}
